backend/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
cached-property==1.5.2
click==8.1.7
exceptiongroup==1.1.3
fastapi==0.104.1
google-auth-oauthlib==0.4.6
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-api-python-client==2.26.1
h11==0.14.0
httplib2==0.20.1
httptools==0.6.1
idna==3.4
inflection==0.5.1
mypy-extensions==1.0.0
mysql-connector-python==8.2.0
Pillow==10.1.0
protobuf==4.21.12
pydantic==2.4.2
pydantic_core==2.10.1
python-dotenv==1.0.0
PyYAML==6.0.1
sniffio==1.3.0
SQLAlchemy==2.0.23
sqlalchemy-orm==1.2.10
starlette==0.27.0
typing==3.7.4.3
typing-inspect==0.9.0
typing_extensions==4.8.0
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
